requirements.txt
Flask==2.1.1
pymongo[srv]==3.12.0
textblob==0.15.3
gunicorn==20.1.0
python-dotenv
Flask==2.1.1
pymongo[srv]==3.12.0
transformers==4.30.2
torch  # Required for Transformers
gunicorn==20.1.0
TextBlob==0.15.3
Flask-Login==0.6.0
Werkzeug==2.0.3
python-dotenv